refactor(#52): move all test cases to e2e package
