+++ src/test/java/com/jrest/java/e2e/RequestBodyTest.java
@@ -1,4 +1,4 @@
-package com.jrest.java;
+package com.jrest.java.e2e;
 
 import com.jrest.java.api.APIResponse;
 import com.jrest.java.api.JRest;
